feat(ritk-filter): SaltAndPepperNoise bit-exact via ITK MT19937 port
Second RNG filter unblocked by porting ITK's exact generator. Add
noise/mersenne.rs: itk::Statistics::MersenneTwisterRandomVariateGenerator
(MT19937) bit-for-bit -- ITK seeding 1812433253*(s^(s>>30))+i, twist
M=397, temper, GetVariate = int/(2^32-1). Rewrote noise/salt_pepper.rs to
ITK semantics: seed = Hash(userSeed,0), per voxel `if GetVariate()<prob {
if GetVariate()<0.5 salt else pepper }`, salt/pepper = +/-f32::MAX (was
image min/max + single draw + StdRng). seed u64 -> u32.

array_equal to sitk.SaltAndPepperNoise single-threaded across 3 prob/seed
cases. Coverage 270 -> 271/298. Shot (Poisson) and Speckle (Normal*input,
FastNorm already ported) now reachable too.

+++ crates/ritk-python/src/filter/noise.rs
@@ -55,12 +55,12 @@ pub fn additive_gaussian_noise(
 /// Returns:
 ///     PyImage with salt-and-pepper noise.
 #[pyfunction]
-#[pyo3(signature = (image, probability, seed=42_u64))]
+#[pyo3(signature = (image, probability, seed=42_u32))]
 pub fn salt_and_pepper_noise(
     py: Python<'_>,
     image: &PyImage,
     probability: f64,
-    seed: u64,
+    seed: u32,
 ) -> RitkResult<PyImage> {
     let img = Arc::clone(&image.inner);
     let result = py
